# swagger/v1/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Makerspace Server V2",
    "version": "v2"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
...
  "paths": {
...
    "/client_error_handler": {
      "post": {
        "summary": "Sends a slack message",
        "tags": [
          "ClientErrorHandler"
        ],
        "operationId": "message",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Message sent"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "messageDetails",
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  }
                },
                "required": [
                  "message"
                ]
              }
            }
          }
        }
      }
    },
...
  }
}